prep

+++ lib/functions/clear_temp_dirs.dart
@@ -3,24 +3,32 @@ import 'dart:io';
 import 'package:pso2_mod_manager/loaders/paths_loader.dart';
 
 void clearAllTempDirs() {
-  Directory(modManAddModsTempDirPath).listSync(recursive: false).forEach((element) {
+  if (Directory(modManAddModsTempDirPath).existsSync()) {
+    Directory(modManAddModsTempDirPath).listSync(recursive: false).forEach((element) {
     if (element.existsSync()) {
       element.deleteSync(recursive: true);
     }
   });
-  Directory(modManAddModsUnpackDirPath).listSync(recursive: false).forEach((element) {
+  }
+  if (Directory(modManAddModsUnpackDirPath).existsSync()) {
+    Directory(modManAddModsUnpackDirPath).listSync(recursive: false).forEach((element) {
     if (element.existsSync()) {
       element.deleteSync(recursive: true);
     }
   });
-  Directory(modManModsAdderPath).listSync(recursive: false).forEach((element) {
+  }
+  if (Directory(modManModsAdderPath).existsSync()) {
+    Directory(modManModsAdderPath).listSync(recursive: false).forEach((element) {
     if (element.existsSync()) {
       element.deleteSync(recursive: true);
     }
   });
-  Directory(modManSwapperDirPath).listSync(recursive: false).forEach((element) {
+  }
+  if (Directory(modManSwapperDirPath).existsSync()) {
+    Directory(modManSwapperDirPath).listSync(recursive: false).forEach((element) {
     if (element.existsSync()) {
       element.deleteSync(recursive: true);
     }
   });
+  }
 }
